+++ src/matchers.ts
@@ -4,6 +4,11 @@ import { DefaultEmbeddingProvider, DefaultGradingProvider } from './providers/op
 import { getNunjucksEngine } from './util';
 import { loadApiProvider } from './providers';
 import {
+  DEFAULT_GRADING_PROMPT,
+  OPENAI_CLOSED_QA_PROMPT,
+  OPENAI_FACTUALITY_PROMPT,
+} from './prompts';
+import {
   ANSWER_RELEVANCY_GENERATE,
   CONTEXT_FAITHFULNESS_LONGFORM,
   CONTEXT_FAITHFULNESS_NLI_STATEMENTS,
@@ -11,10 +16,7 @@ import {
   CONTEXT_RECALL_ATTRIBUTED_TOKEN,
   CONTEXT_RELEVANCE,
   CONTEXT_RELEVANCE_BAD,
-  DEFAULT_GRADING_PROMPT,
-  OPENAI_CLOSED_QA_PROMPT,
-  OPENAI_FACTUALITY_PROMPT,
-} from './prompts';
+} from './external/ragas';
 
 import type {
   ApiClassificationProvider,
@@ -500,7 +502,6 @@ export async function matchesAnswerRelevance(
   };
 }
 
-// Ported from RAGAS. https://github.com/explodinggradients/ragas. See APACHE_LICENSE for license.
 export async function matchesContextRecall(
   context: string,
   groundTruth: string,
@@ -546,7 +547,6 @@ export async function matchesContextRecall(
   };
 }
 
-// Ported from RAGAS. https://github.com/explodinggradients/ragas. See APACHE_LICENSE for license.
 export async function matchesContextRelevance(
   question: string,
   context: string,
@@ -590,7 +590,6 @@ export async function matchesContextRelevance(
   };
 }
 
-// Ported from RAGAS. https://github.com/explodinggradients/ragas. See APACHE_LICENSE for license.
 export async function matchesContextFaithfulness(
   query: string,
   output: string,
